feat: user email + sign-out on chat builder header
Admin layout has shown email + sign-out since M2c-2; the chat builder
at / had no user indicator and no way to sign out without first
navigating into /admin/*. Parity fix:

- lib/current-user.ts (new) — resolveCurrentUser helper that walks
  the same flag / kill-switch / getCurrentUser path as lib/admin-gate
  but returns null (instead of driving a redirect) when no identity
  is available. Intended for non-gated surfaces that want to render
  the current user when signed in and otherwise show nothing.

- app/page.tsx — split into a server shell. Resolves the current
  user and passes email down to the client component. force-dynamic
  so the page reads cookies per request (was implicitly dynamic
  already via the "use client" subtree, but being explicit prevents
  future accidental static optimisation).

- components/HomePageClient.tsx — the old app/page.tsx client body,
  now receives userEmail: string | null. Header gains a right-hand
  strip matching the admin layout: email, Admin → link, Sign out
  form posting to /logout. Strip hides when userEmail is null
  (flag off / kill switch / no session).

Verification: tsc / lint / build clean; / flipped from ○ (static)
to ƒ (dynamic) in the build report, which is expected.
